refactor(gazetteer): extract ingest-wof / fold-overture / fold-geonames / freeze into the pipeline module
build-unified-wof.ts is now a thin CLI over mailwoman/gazetteer-pipeline (exported at ./gazetteer-pipeline). fold-geonames drops the hardcoded playpen dump dir for dataRootPath. freezeAdmin carries the #1015 ordering fixes (ancestors_by_id before the -4 backfill, maxId bound) with a fixture test.
