+++ scripts/build-unified-wof.ts
@@ -3,109 +3,44 @@
  * @license AGPL-3.0
  * @author Teffen Ellis, et al.
  *
- *   Build a unified WOF SQLite database from cloned GeoJSON repos. Implements the WAL + Freeze design
- *   brief (docs/articles/reviews/2026-05-28-sqlite-wal-strategy.md).
+ *   Build a unified WOF SQLite database from cloned GeoJSON repos — now a thin CLI over the
+ *   `mailwoman/gazetteer-pipeline` module (ingest-wof → fold-overture → fold-geonames → freeze),
+ *   which is where the implementation lives. SUPERSEDED by `mailwoman gazetteer build admin`
+ *   (which adds the enrich/FTS/verify/seal steps this script never had); kept only until the
+ *   command lands — see docs/superpowers/specs/2026-07-07-scripts-cleanup-gazetteer-cli-design.md.
  *
- *   Phase 1: Enumerate GeoJSON files across one or more repo directories. Phase 2: Ingest — parallel
- *   file reads (asyncParallelIterator), single-thread writer, WAL mode. Phase 2b (optional):
- *   backfill the Overture `divisions` theme for locales the WOF repos don't cover
- *   (`--overture-countries`). Phase 3: Freeze — checkpoint, journal_mode DELETE, indexes, ANALYZE,
- *   VACUUM INTO.
- *
- *   Usage: node scripts/build-unified-wof.js\
- *   --data /mnt/playpen/mailwoman-data/wof/repos/whosonfirst-data\
- *   --output /mnt/playpen/mailwoman-data/wof/admin-global.db\
- *   [--concurrency 64] [--batch 500] [--overture-countries PT,PL,CZ,NO,FI,HR]
- *
- *   Accepts a parent directory containing multiple whosonfirst-data-admin-* subdirectories, or a
- *   single repo directory. `--overture-countries` folds the zero-DB-locale coverage (proven in the
- *   2026-06-20 sprint, project-eu-coverage-not-retrain) into the canonical build so the shipped
- *   admin DB carries global admin coverage in one file — superseding the standalone
- *   scripts/build-overture-divisions-gazetteer.py prototype. The Overture FTS rides the same
- *   `build-fts` step as the WOF rows.
+ *   Usage: node scripts/build-unified-wof.ts\
+ *   --data /mnt/playpen/mailwoman-data/wof/repos\
+ *   --output <staging.db>\
+ *   [--overture-countries …] [--geonames-countries …] [--overture-release 2026-06-17.0]
  */
 
 import { existsSync, statSync, unlinkSync } from "node:fs"
-import { readFile } from "node:fs/promises"
 import { DatabaseSync } from "node:sqlite"
 
-import { DuckDBInstance } from "@duckdb/node-api"
-import { isOfficialLanguage } from "@mailwoman/codex/country"
 import { dataRootPath, sealDatabase } from "@mailwoman/core/utils"
+import { createUnifiedSchema } from "@mailwoman/resolver-wof-sqlite/unified-schema"
 import {
-	backfillAncestorsFromHierarchy,
-	discoverAdminDataRoots,
-} from "@mailwoman/resolver-wof-sqlite/ancestry-backfill"
-import { buildCoincidentRoles } from "@mailwoman/resolver-wof-sqlite/coincident-roles"
-import { ingestGeonamesAliases } from "@mailwoman/resolver-wof-sqlite/geonames-aliases"
-import { ingestGeonamesPostal } from "@mailwoman/resolver-wof-sqlite/geonames-postal"
-import {
-	createUnifiedIndexes,
-	createUnifiedSchema,
-	populateAncestors,
-} from "@mailwoman/resolver-wof-sqlite/unified-schema"
-import FastGlob from "fast-glob"
-import { asyncParallelIterator } from "spliterator"
-
-/**
- * Synthetic id base for Overture-sourced rows — above any real WOF id (WOF ids are <~2e9), so a combined DB never
- * collides across sources. Matches scripts/build-overture-divisions-gazetteer.py (the standalone falsification
- * prototype this folds into the unified build).
- */
-const OVERTURE_ID_BASE = 8_000_000_000_000
-/**
- * Overture division subtypes that map to the resolver's admin placetypes. `country` is included (#1015) so an
- * Overture-backfilled locale gets its country node — without it the reverse geocoder has no country tier to anchor to
- * (Brussels → nearest FOREIGN place across the border), and forward resolution can't country-gate the locale.
- */
-const OVERTURE_DIVISION_SUBTYPES = ["country", "locality", "region", "county", "localadmin"]
-/** Default GeoNames per-country dump directory (`<CC>.txt`, from download.geonames.org/export/dump). */
-const DEFAULT_GEONAMES_DIR = "/mnt/playpen/mailwoman-data/geonames"
-/** Default GeoNames per-country POSTAL dump directory (`<CC>.txt`, from download.geonames.org/export/zip). */
-const DEFAULT_GEONAMES_POSTAL_DIR = String(dataRootPath("geonames-postal"))
-/** Default GeoNames per-country alternateNamesV2 directory (`<CC>.txt`, from …/export/dump/alternatenames). */
-const DEFAULT_GEONAMES_ALTERNATE_DIR = String(dataRootPath("geonames-alternate"))
-/** Pinned Overture release for the divisions theme (the release the EU coverage was validated on). */
-const DEFAULT_OVERTURE_RELEASE = "2026-06-17.0"
+	DEFAULT_OVERTURE_RELEASE,
+	foldGeonames,
+	freezeAdmin,
+	ingestOvertureDivisions,
+	ingestWOF,
+} from "mailwoman/gazetteer-pipeline"
 
 interface Args {
 	dataDir: string
 	outputPath: string
 	concurrency: number
 	batchCommitSize: number
-	/**
-	 * Override the ingested placetype set (comma-separated). Defaults to ADMIN_PLACETYPES; pass `--placetypes postalcode`
-	 * to build the postcode shard from whosonfirst-data-postalcode-* repos.
-	 */
+	/** Override the ingested placetype set; pass `--placetypes postalcode` for the postcode shard. */
 	placetypes?: string[]
-	/**
-	 * Comma-separated ISO 3166-1 alpha-2 codes to backfill from the Overture `divisions` theme AFTER the WOF GeoJSON
-	 * ingest — the zero-DB locales the WOF repos don't cover (PT/PL/CZ/NO/FI/HR/…). Rows land in the SAME
-	 * spr/names/place_population tables with synthetic ids, so the Freeze phase indexes them uniformly. Off unless
-	 * provided. See project-eu-coverage-not-retrain.
-	 */
 	overtureCountries?: string[]
-	/** Overture release for `--overture-countries` (the divisions theme is pinned per release). */
 	overtureRelease: string
-	/**
-	 * #743/#193: comma-separated ISO codes to backfill GeoNames place ALIASES (the bilingual / alt-name class —
-	 * Karjaa↔Karis) AFTER the WOF + Overture ingest. Reads `<CC>.txt` from {@link geonamesDir}. Off unless provided.
-	 */
 	geonamesCountries?: string[]
-	/**
-	 * #920: fold GeoNames POSTAL codes (download.geonames.org/export/zip) for these countries as `postalcode` places —
-	 * the tail locales without WOF postalcode repos. Name law + medoid centroids enforced in
-	 * `@mailwoman/resolver-wof-sqlite/geonames-postal`. Off unless provided.
-	 */
 	geonamesPostalCountries?: string[]
 	geonamesPostalDir: string
-	/** Directory of GeoNames per-country dumps (`<CC>.txt`). Default {@link DEFAULT_GEONAMES_DIR}. */
 	geonamesDir: string
-	/**
-	 * #936: directory of GeoNames per-country alternateNamesV2 dumps
-	 * (`download.geonames.org/export/dump/alternatenames/<CC>.zip` → `<CC>.txt`) — decorates the `--geonames-countries`
-	 * alias rows with language / privateuse / the `official` bit. Countries without a file ingest untagged (not fatal).
-	 */
 	geonamesAlternateDir: string
 }
 
@@ -119,10 +54,10 @@ function parseArgs(): Args {
 	let overtureCountries: string[] | undefined
 	let overtureRelease = DEFAULT_OVERTURE_RELEASE
 	let geonamesCountries: string[] | undefined
-	let geonamesDir = DEFAULT_GEONAMES_DIR
+	let geonamesDir = String(dataRootPath("geonames"))
 	let geonamesPostalCountries: string[] | undefined
-	let geonamesPostalDir = DEFAULT_GEONAMES_POSTAL_DIR
-	let geonamesAlternateDir = DEFAULT_GEONAMES_ALTERNATE_DIR
+	let geonamesPostalDir = String(dataRootPath("geonames-postal"))
+	let geonamesAlternateDir = String(dataRootPath("geonames-alternate"))
 
 	for (let i = 0; i < args.length; i++) {
 		if (args[i] === "--data" && args[i + 1]) dataDir = args[++i]
@@ -150,7 +85,7 @@ function parseArgs(): Args {
 
 	if (!dataDir || !outputPath) {
 		console.error(
-			"Usage: node scripts/build-unified-wof.js --data <wof-repos-dir> --output <output.db> [--concurrency 64] [--batch 500] [--placetypes postalcode] [--overture-countries PT,PL,CZ] [--overture-release 2026-06-17.0] [--geonames-countries FI,EE,LV] [--geonames-dir <dir>]"
+			"Usage: node scripts/build-unified-wof.ts --data <wof-repos-dir> --output <output.db> [--concurrency 64] [--batch 500] [--placetypes postalcode] [--overture-countries PT,PL,CZ] [--overture-release 2026-06-17.0] [--geonames-countries FI,EE,LV] [--geonames-dir <dir>]  — SUPERSEDED by `mailwoman gazetteer build admin`"
 		)
 		process.exit(1)
 	}
@@ -171,315 +106,13 @@ function parseArgs(): Args {
 	}
 }
 
-const ADMIN_PLACETYPES = new Set([
-	"country",
-	"region",
-	"county",
-	"locality",
-	"localadmin",
-	"borough",
-	"neighbourhood",
-	"macroregion",
-	"macrocounty",
-])
-
-interface ParsedFeature {
-	id: number
-	parent_id: number
-	name: string
-	placetype: string
-	country: string
-	latitude: number
-	longitude: number
-	minLatitude: number
-	minLongitude: number
-	maxLatitude: number
-	maxLongitude: number
-	population: number
-	isCurrent: number
-	isDeprecated: number
-	isCeased: number
-	isSuperseded: number
-	isSuperseding: number
-	lastmodified: number
-	concordances: Record<string, string | number>
-	names: Array<{ name: string; language: string; privateuse: string; official: number }>
-}
-
-function parseFeature(text: string, placetypes: Set<string>): ParsedFeature | null {
-	const feature = JSON.parse(text)
-	const props = feature.properties
-
-	if (!props) return null
-
-	const supersededBy = props["wof:superseded_by"]
-
-	if (supersededBy && supersededBy.length > 0) return null
-
-	const placetype = props["wof:placetype"]
-
-	if (!placetypes.has(placetype)) return null
-
-	const mzIsCurrent = props["mz:is_current"]
-
-	const lat = typeof props["geom:latitude"] === "number" ? props["geom:latitude"] : 0
-	const lon = typeof props["geom:longitude"] === "number" ? props["geom:longitude"] : 0
-	// WOF `geom:bbox` is "minLon,minLat,maxLon,maxLat". Fall back to the centroid (a point bbox) when
-	// absent — still correct for point-in-box proximity, the resolver's main bbox use.
-	let [minLon, minLat, maxLon, maxLat] = [lon, lat, lon, lat]
-	const bboxStr = props["geom:bbox"]
-
-	if (typeof bboxStr === "string") {
-		const parts = bboxStr.split(",").map(Number)
-
-		if (parts.length === 4 && parts.every((n) => Number.isFinite(n))) {
-			;[minLon, minLat, maxLon, maxLat] = parts as [number, number, number, number]
-		}
-	}
-
-	const names: Array<{ name: string; language: string; privateuse: string; official: number }> = []
-	const country = props["wof:country"] ?? ""
-
-	for (const [key, value] of Object.entries(props)) {
-		const match = key.match(/^name:([a-z]{3})_x_(preferred|variant)$/)
-
-		if (!match || !value) continue
-		const lang = match[1]!
-		const privateuse = match[2]!
-		// #936: only PREFERRED forms in an official language are official names — x_variant rows
-		// tagged with an official language are abbreviations/codes ("MSP", "Frisco"), and marking
-		// them official scored 13× the collision count in the risk probe.
-		const official = privateuse === "preferred" && isOfficialLanguage(country, lang) ? 1 : 0
-		const vals = Array.isArray(value) ? value : [value]
-
-		for (const v of vals) {
-			if (typeof v === "string" && v.length > 0) {
-				names.push({ name: v, language: lang, privateuse, official })
-			}
-		}
-	}
-
-	return {
-		id: props["wof:id"],
-		parent_id: props["wof:parent_id"] ?? -1,
-		name: props["wof:name"] ?? "",
-		placetype,
-		country: props["wof:country"] ?? "",
-		latitude: lat,
-		longitude: lon,
-		minLatitude: minLat,
-		minLongitude: minLon,
-		maxLatitude: maxLat,
-		maxLongitude: maxLon,
-		population: props["wof:population"] ?? props["gn:population"] ?? 0,
-		isCurrent: mzIsCurrent === 0 || mzIsCurrent === "0" ? 0 : 1,
-		isDeprecated: props["edtf:deprecated"] ? 1 : 0,
-		isCeased: props["edtf:cessation"] ? 1 : 0,
-		isSuperseded: (props["wof:superseded_by"]?.length ?? 0) > 0 ? 1 : 0,
-		isSuperseding: (props["wof:supersedes"]?.length ?? 0) > 0 ? 1 : 0,
-		lastmodified: typeof props["wof:lastmodified"] === "number" ? props["wof:lastmodified"] : 0,
-		concordances: props["wof:concordances"] ?? {},
-		names,
-	}
-}
-
-/**
- * Backfill the Overture `divisions` theme into an already-open unified ingest DB, for locales the WOF GeoJSON repos
- * don't cover (the 2026-06-20 zero-DB EU set). Writes the SAME spr/names/ place_population tables the WOF path uses —
- * with synthetic ids based at {@link OVERTURE_ID_BASE} so the two sources never collide — so the caller's Freeze phase
- * (ancestors closure, coincident_roles, indexes, FTS) treats them uniformly. The Overture sub-tree is self-contained
- * (locality → region → county via `parent_division_id`); a division whose parent we didn't ingest tops out at -1.
- * Country scoping rides `spr.country` (set on every row), not the ancestry, so no WOF country node is needed.
- *
- * The heavy native `@duckdb/node-api` dependency lives here in `scripts/` only — never in `@mailwoman/corpus` (a
- * runtime dep of the `mailwoman` CLI) — the same split as `ingest-overture-addresses.ts`.
- *
- * @returns The number of divisions ingested.
- */
-export async function ingestOvertureDivisions(
-	db: DatabaseSync,
-	countries: string[],
-	release: string,
-	/**
-	 * Starting synthetic id. Defaults to {@link OVERTURE_ID_BASE} (a single full build). An INCREMENTAL augment of a DB
-	 * that ALREADY holds Overture rows MUST pass `max(spr.id) + 1` so the new ids don't collide with — and `INSERT OR
-	 * REPLACE` clobber — the existing ones.
-	 */
-	idBase: number = OVERTURE_ID_BASE
-): Promise<number> {
-	const inlist = countries.map((c) => `'${c.replace(/'/g, "''")}'`).join(",")
-	const subtypes = OVERTURE_DIVISION_SUBTYPES.map((s) => `'${s}'`).join(",")
-	const glob = `s3://overturemaps-us-west-2/release/${release}/theme=divisions/type=division/*`
-	// #1015: the real boundary EXTENT lives in the sibling `type=division_area` (the polygon). The `type=division`
-	// row is the label POINT — its `bbox` is a degenerate point, so relying on it left every Overture-backfilled
-	// place with a point bbox, invisible to the reverse geocoder's bbox-containment (Brussels resolved to a foreign
-	// cross-border neighbour). Join the area's extent by `division_id`, falling back to the point bbox when a
-	// division has no area row (so nothing regresses).
-	const areaGlob = `s3://overturemaps-us-west-2/release/${release}/theme=divisions/type=division_area/*`
-
-	const instance = await DuckDBInstance.create()
-	const con = await instance.connect()
-	await con.run(
-		"INSTALL httpfs; LOAD httpfs; INSTALL spatial; LOAD spatial; INSTALL json; LOAD json; SET s3_region='us-west-2';"
-	)
-	await con.run("SET memory_limit='4GB'; SET threads=4;")
-
-	console.error(`  Overture divisions: querying ${countries.join(",")} @ release ${release}...`)
-	const result = await con.runAndReadAll(`
-		WITH area AS (
-			SELECT division_id,
-				MIN(bbox.ymin) AS ymin, MAX(bbox.ymax) AS ymax, MIN(bbox.xmin) AS xmin, MAX(bbox.xmax) AS xmax
-			FROM read_parquet('${areaGlob}')
-			WHERE country IN (${inlist})
-			GROUP BY division_id
-		)
-		SELECT d.id AS id,
-			d.names.primary AS name,
-			to_json(d.names.common) AS common_json,
-			d.subtype AS subtype,
-			d.country AS country,
-			ST_Y(ST_Centroid(d.geometry)) AS lat,
-			ST_X(ST_Centroid(d.geometry)) AS lon,
-			COALESCE(a.ymin, d.bbox.ymin) AS min_lat, COALESCE(a.ymax, d.bbox.ymax) AS max_lat,
-			COALESCE(a.xmin, d.bbox.xmin) AS min_lon, COALESCE(a.xmax, d.bbox.xmax) AS max_lon,
-			d.parent_division_id AS parent_division_id,
-			d.population AS population
-		FROM read_parquet('${glob}') d
-		LEFT JOIN area a ON a.division_id = d.id
-		WHERE d.country IN (${inlist}) AND d.subtype IN (${subtypes})
-			AND d.names.primary IS NOT NULL AND d.geometry IS NOT NULL
-	`)
-	const rows = result.getRowObjects() as Array<Record<string, unknown>>
-	console.error(`  Overture divisions: ${rows.length.toLocaleString()} pulled`)
-
-	// GERS string id → synthetic int, sequential and unique within this run.
-	const idmap = new Map<string, number>()
-	rows.forEach((r, i) => idmap.set(String(r.id), idBase + i))
-
-	const sprInsert = db.prepare(
-		`INSERT OR REPLACE INTO spr (id, parent_id, name, placetype, country, latitude, longitude, min_latitude, min_longitude, max_latitude, max_longitude, is_current, is_deprecated, is_ceased, is_superseded, is_superseding, lastmodified) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)`
-	)
-	const namesInsert = db.prepare(
-		`INSERT INTO names (id, name, placetype, country, language, official, lastmodified) VALUES (?, ?, ?, ?, ?, ?, ?)`
-	)
-	const populationInsert = db.prepare(`INSERT OR REPLACE INTO place_population (id, population) VALUES (?, ?)`)
-
-	const num = (v: unknown): number => (typeof v === "number" ? v : typeof v === "bigint" ? Number(v) : 0)
-	// Keep only Latin-script common-name aliases (English + major-language transliterations — the names a
-	// Latin-keyboard user actually queries: "Moscow", "Moscou", "Moskva"). The local-script primary
-	// (Москва, القاهرة) is kept separately; obscure non-Latin aliases (Armenian, Mingrelian, …) would
-	// bloat the candidate for ~zero query value.
-	const isLatin = (s: string): boolean => /^[\p{Script=Latin}\p{N}\p{P}\s]+$/u.test(s)
-
-	db.exec("BEGIN")
-	let n = 0
-
-	for (const r of rows) {
-		const nid = idmap.get(String(r.id))!
-		const pgers = r.parent_division_id == null ? null : String(r.parent_division_id)
-		const pid = (pgers && idmap.get(pgers)) || -1
-		const name = String(r.name)
-		const subtype = String(r.subtype)
-		const country = String(r.country ?? "").toUpperCase()
-		// SELECT aliases: min_lat=ymin, min_lon=xmin, max_lat=ymax, max_lon=xmax → spr (lat, lon,
-		// min_latitude, min_longitude, max_latitude, max_longitude).
-		sprInsert.run(
-			nid,
-			pid,
-			name,
-			subtype,
-			country,
-			num(r.lat),
-			num(r.lon),
-			num(r.min_lat),
-			num(r.min_lon),
-			num(r.max_lat),
-			num(r.max_lon),
-			1,
-			0,
-			0,
-			0,
-			0,
-			0
-		)
-		namesInsert.run(nid, name, subtype, country, "", 0, 0)
-
-		// Multilingual aliases (names.common — language→name, incl. English / Latin transliterations) so a
-		// non-Latin-script place (Москва, القاهرة, กรุงเทพมหานคร) still resolves by its English/Latin name.
-		// The candidate build explodes every alias here into its own name_key. Overture `common` is the
-		// standard name per language (no variant axis), so #936 officialness is the language test alone.
-		if (r.common_json) {
-			try {
-				const common = JSON.parse(String(r.common_json)) as Record<string, string>
-				const seen = new Set([name])
-
-				for (const [lang, alias] of Object.entries(common)) {
-					if (typeof alias === "string" && alias.length > 0 && !seen.has(alias) && isLatin(alias)) {
-						seen.add(alias)
-						namesInsert.run(nid, alias, subtype, country, lang, isOfficialLanguage(country, lang) ? 1 : 0, 0)
-					}
-				}
-			} catch {
-				/* malformed common map — keep the primary, skip aliases */
-			}
-		}
-		const pop = num(r.population)
-
-		if (pop > 0) populationInsert.run(nid, pop)
-		n++
-	}
-	db.exec("COMMIT")
-
-	return n
-}
-
-// ingestGeonamesAliases moved to @mailwoman/resolver-wof-sqlite/geonames-aliases — ONE shared
-// implementation across this canonical `--geonames-countries` build, the standalone
-// `build-admin-geonames-fold`, and the `mailwoman gazetteer` commands. Imported at the top.
-
 async function main() {
-	const {
-		dataDir,
-		outputPath,
-		concurrency,
-		batchCommitSize,
-		placetypes,
-		overtureCountries,
-		overtureRelease,
-		geonamesCountries,
-		geonamesDir,
-		geonamesPostalCountries,
-		geonamesPostalDir,
-		geonamesAlternateDir,
-	} = parseArgs()
-	const activePlacetypes = placetypes ? new Set(placetypes) : ADMIN_PLACETYPES
-	console.error(`Ingesting placetypes: ${[...activePlacetypes].join(", ")}`)
+	const args = parseArgs()
 	const t0 = performance.now()
-	const ingestPath = outputPath + ".ingest"
+	const ingestPath = args.outputPath + ".ingest"
 
 	if (existsSync(ingestPath)) unlinkSync(ingestPath)
 
-	// -----------------------------------------------------------------------
-	// Phase 1: Enumerate
-	// -----------------------------------------------------------------------
-	console.error(`Scanning ${dataDir} for GeoJSON files...`)
-	// The whosonfirst-data-postalcode-* repos hold millions of postcode geojson the ADMIN build filters out by
-	// placetype anyway — enumerating + reading them is the bulk of the ingest time. Exclude them unless this build
-	// actually wants postalcode rows (#1015 — a full admin rebuild off `--data …/repos` dropped from a ~30-min scan of
-	// the postcode tail to a couple of minutes).
-	const ignore = ["**/*-alt-*"]
-
-	if (!activePlacetypes.has("postalcode")) ignore.push("**/whosonfirst-data-postalcode-*/**")
-	const filePaths = await FastGlob("**/data/**/*.geojson", {
-		cwd: dataDir,
-		absolute: true,
-		ignore,
-	})
-	console.error(`  Found ${filePaths.length.toLocaleString()} files (excluding -alt- variants)`)
-
-	// -----------------------------------------------------------------------
-	// Phase 2: Ingest (WAL mode, single-thread writer, parallel file reads)
-	// -----------------------------------------------------------------------
 	console.error(`Creating ingest DB at ${ingestPath}...`)
 	const db = new DatabaseSync(ingestPath)
 	db.exec(`
@@ -492,226 +125,61 @@ async function main() {
 	`)
 	await createUnifiedSchema(db)
 
-	const sprInsert = db.prepare(
-		`INSERT OR REPLACE INTO spr (id, parent_id, name, placetype, country, latitude, longitude, min_latitude, min_longitude, max_latitude, max_longitude, is_current, is_deprecated, is_ceased, is_superseded, is_superseding, lastmodified) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)`
-	)
-	const namesInsert = db.prepare(
-		`INSERT INTO names (id, name, placetype, country, language, privateuse, official, lastmodified) VALUES (?, ?, ?, ?, ?, ?, ?, ?)`
-	)
-	const concordancesInsert = db.prepare(
-		`INSERT INTO concordances (id, other_id, other_source, lastmodified) VALUES (?, ?, ?, ?)`
-	)
-	const populationInsert = db.prepare(`INSERT OR REPLACE INTO place_population (id, population) VALUES (?, ?)`)
-
-	let processed = 0
-	let skipped = 0
-	let inTransaction = false
-
-	function beginIfNeeded() {
-		if (!inTransaction) {
-			db.exec("BEGIN TRANSACTION")
-			inTransaction = true
-		}
-	}
-
-	function commitIfNeeded(force = false) {
-		if (inTransaction && (force || processed % batchCommitSize === 0)) {
-			db.exec("COMMIT")
-			inTransaction = false
-		}
-	}
-
-	console.error(`Ingesting with concurrency=${concurrency}, batch commit every ${batchCommitSize} files...`)
-
-	const readResults = asyncParallelIterator(filePaths, concurrency, (filePath) => readFile(filePath, "utf8"))
-
-	for await (const text of readResults) {
-		const feature = parseFeature(text, activePlacetypes)
-
-		if (!feature) {
-			skipped++
-			continue
-		}
-
-		beginIfNeeded()
-
-		sprInsert.run(
-			feature.id,
-			feature.parent_id,
-			feature.name,
-			feature.placetype,
-			feature.country,
-			feature.latitude,
-			feature.longitude,
-			feature.minLatitude,
-			feature.minLongitude,
-			feature.maxLatitude,
-			feature.maxLongitude,
-			feature.isCurrent,
-			feature.isDeprecated,
-			feature.isCeased,
-			feature.isSuperseded,
-			feature.isSuperseding,
-			feature.lastmodified
-		)
-
-		for (const n of feature.names) {
-			namesInsert.run(
-				feature.id,
-				n.name,
-				feature.placetype,
-				feature.country,
-				n.language,
-				n.privateuse,
-				n.official,
-				feature.lastmodified
-			)
-		}
-
-		for (const [source, value] of Object.entries(feature.concordances)) {
-			concordancesInsert.run(feature.id, String(value), source, feature.lastmodified)
-		}
-
-		if (feature.population > 0) {
-			populationInsert.run(feature.id, feature.population)
-		}
-
-		processed++
-		commitIfNeeded()
-
-		if (processed % 25000 === 0) {
+	console.error(`Scanning + ingesting WOF GeoJSON from ${args.dataDir}...`)
+	const ingest = await ingestWOF(db, {
+		dataDir: args.dataDir,
+		placetypes: args.placetypes ? new Set(args.placetypes) : undefined,
+		concurrency: args.concurrency,
+		batchCommitSize: args.batchCommitSize,
+		onProgress: (processed, skipped, total) => {
 			const elapsed = (performance.now() - t0) / 1000
 			const rate = processed / elapsed
-			const eta = (filePaths.length - processed - skipped) / rate
+			const eta = (total - processed - skipped) / rate
 			console.error(
 				`  ${processed.toLocaleString()} processed, ${skipped.toLocaleString()} skipped (${rate.toFixed(0)}/s, ETA ${eta.toFixed(0)}s)`
 			)
-		}
-	}
-
-	commitIfNeeded(true)
-	const ingestElapsed = ((performance.now() - t0) / 1000).toFixed(1)
+		},
+	})
 	console.error(
-		`Ingest complete in ${ingestElapsed}s: ${processed.toLocaleString()} places, ${skipped.toLocaleString()} skipped`
+		`Ingest complete: ${ingest.placesIngested.toLocaleString()} places, ${ingest.skipped.toLocaleString()} skipped`
 	)
 
-	// -----------------------------------------------------------------------
-	// Phase 2b: Overture divisions backfill (zero-DB locales the WOF repos miss)
-	// -----------------------------------------------------------------------
 	let overtureIngested = 0
 
-	if (overtureCountries && overtureCountries.length > 0) {
-		console.error(`Backfilling Overture divisions for ${overtureCountries.join(",")}...`)
-		overtureIngested = await ingestOvertureDivisions(db, overtureCountries, overtureRelease)
+	if (args.overtureCountries && args.overtureCountries.length > 0) {
+		console.error(`Backfilling Overture divisions for ${args.overtureCountries.join(",")}...`)
+		overtureIngested = await ingestOvertureDivisions(db, args.overtureCountries, args.overtureRelease)
 		console.error(`  Overture divisions ingested: ${overtureIngested.toLocaleString()}`)
 	}
 
-	// -----------------------------------------------------------------------
-	// Phase 2c: GeoNames bilingual / alt-name backfill (#743/#193 — the Karjaa↔Karis class)
-	// -----------------------------------------------------------------------
-	let geonamesIngested = 0
-
-	if (geonamesCountries && geonamesCountries.length > 0) {
-		console.error(`Backfilling GeoNames aliases for ${geonamesCountries.join(",")} from ${geonamesDir}...`)
-		geonamesIngested = ingestGeonamesAliases(db, geonamesCountries, geonamesDir, undefined, {
-			alternateDir: geonamesAlternateDir,
+	if (args.geonamesCountries && args.geonamesCountries.length > 0) {
+		console.error(`Backfilling GeoNames aliases for ${args.geonamesCountries.join(",")}...`)
+		const folded = foldGeonames(db, {
+			countries: args.geonamesCountries,
+			geonamesDir: args.geonamesDir,
+			alternateDir: args.geonamesAlternateDir,
+			postalCountries: args.geonamesPostalCountries,
+			postalDir: args.geonamesPostalDir,
 		})
-		console.error(`  GeoNames places ingested: ${geonamesIngested.toLocaleString()}`)
-	}
-
-	// -----------------------------------------------------------------------
-	// Phase 2d: GeoNames postal backfill (#920 — postcode coverage for the namesake-tail locales)
-	// -----------------------------------------------------------------------
-	if (geonamesPostalCountries && geonamesPostalCountries.length > 0) {
 		console.error(
-			`Backfilling GeoNames postal codes for ${geonamesPostalCountries.join(",")} from ${geonamesPostalDir}...`
+			`  GeoNames places ingested: ${folded.placesIngested.toLocaleString()} (+${folded.postalIngested.toLocaleString()} postal)`
 		)
-		const postal = ingestGeonamesPostal(db, geonamesPostalCountries, geonamesPostalDir)
-		console.error(`  GeoNames postal codes ingested: ${postal.inserted.toLocaleString()}`)
 	}
 
-	// -----------------------------------------------------------------------
-	// Phase 3: Freeze
-	// -----------------------------------------------------------------------
 	console.error("Freezing...")
+	const frozen = await freezeAdmin(db, {
+		dataDir: args.dataDir,
+		onPhase: (phase, detail) => console.error(`  [${phase}]${detail ? ` ${detail}` : ""}`),
+	})
+	console.error(`  ancestors: ${frozen.ancestorRows} rows; coincident_roles: ${frozen.coincidentRoles}`)
 
-	const checkpoint = db.prepare("PRAGMA wal_checkpoint(TRUNCATE)").get() as {
-		busy: number
-		log: number
-		checkpointed: number
+	if (existsSync(args.outputPath)) {
+		const s = statSync(args.outputPath)
+
+		if (s.size > 0) unlinkSync(args.outputPath)
 	}
-
-	if (checkpoint.busy !== 0) {
-		throw new Error(`WAL checkpoint did not finish: ${JSON.stringify(checkpoint)}`)
-	}
-	console.error("  WAL checkpoint complete")
-
-	const mode = db.prepare("PRAGMA journal_mode = DELETE").get() as { journal_mode: string }
-
-	if (mode.journal_mode !== "delete") {
-		throw new Error(`journal_mode switch failed; still ${mode.journal_mode}`)
-	}
-	console.error("  journal_mode = delete")
-
-	console.error("  Building ancestors (parent_id closure)...")
-	const ancestorRows = populateAncestors(db)
-	console.error(`  ancestors: ${ancestorRows} rows`)
-
-	// Index `ancestors(id)` NOW — before the −4 backfill probes it. `createUnifiedIndexes` (below) builds this same
-	// index, but it runs AFTER the backfill; without it here the backfill's per-candidate `count(*)/EXISTS` lookups
-	// full-scan the 13M-row `ancestors` table EACH time, and the build stalls for hours (found rebuilding for #1015).
-	// `ifNotExists` keeps the later `createUnifiedIndexes` a no-op.
-	console.error("  Indexing ancestors(id) for the backfill...")
-	db.exec("CREATE INDEX IF NOT EXISTS ancestors_by_id ON ancestors(id)")
-
-	// Repair the multi-parent orphans the closure leaves only-self (#440 / #832): places with
-	// `wof:parent_id = -4` (New York City, London, …) get NO region/county ancestry from the closure,
-	// so the resolver's region-descendant filter excludes them and a small namesake wins. Read the
-	// authoritative `wof:hierarchy` from source geojson and insert the missing rows. MUST run here —
-	// after the closure, before `coincident_roles` (which reads ancestry) and the VACUUM freeze.
-	console.error("  Backfilling ancestry from wof:hierarchy (multi-parent -4 places)...")
-	const geojsonRoots = discoverAdminDataRoots(dataDir)
-
-	if (geojsonRoots.length === 0) {
-		console.error(`    WARNING: no */data geojson roots under ${dataDir}; skipping — orphans like NYC stay unreachable`)
-	} else {
-		// Only real WOF places have `wof:hierarchy` geojson; synthetic Overture/GeoNames rows (ids >= OVERTURE_ID_BASE)
-		// never do, and probing millions of them across every repo root is what turned this step into a ~40-min stall on
-		// the wide-coverage build. Their ancestry comes from the parent_id closure above, not this backfill.
-		const bf = backfillAncestorsFromHierarchy(db, geojsonRoots, { maxId: OVERTURE_ID_BASE })
-		console.error(
-			`    +${bf.rowsAdded} rows for ${bf.placesFixed} places (${bf.noGeojson} candidates had no source geojson)`
-		)
-	}
-
-	// Dual-role-place relation (#403, epic #402) — needs `ancestors` + `spr` bbox + `place_population`,
-	// all present by now. Drives the resolver's hierarchy completion (on by default). Tiny (~hundreds of
-	// rows); `build-slim` carries it through to the shipped DB.
-	console.error("  Building coincident_roles (dual-role places)...")
-	const roles = buildCoincidentRoles(db)
-	console.error(`  coincident_roles: ${roles.rowCount} rows`)
-
-	console.error("  Creating indexes...")
-	await createUnifiedIndexes(db)
-
-	console.error("  ANALYZE...")
-	db.exec("ANALYZE")
-	db.exec("PRAGMA optimize")
-
-	const integrity = db.prepare("PRAGMA integrity_check").get() as { integrity_check: string }
-
-	if (integrity.integrity_check !== "ok") {
-		throw new Error(`integrity_check failed: ${integrity.integrity_check}`)
-	}
-	console.error("  integrity_check = ok")
-
-	if (existsSync(outputPath)) {
-		const s = statSync(outputPath)
-
-		if (s.size > 0) unlinkSync(outputPath)
-	}
-	db.prepare("VACUUM INTO ?").run(outputPath)
-	console.error(`  VACUUM INTO ${outputPath}`)
+	db.prepare("VACUUM INTO ?").run(args.outputPath)
+	console.error(`  VACUUM INTO ${args.outputPath}`)
 
 	db.close()
 	unlinkSync(ingestPath)
@@ -720,31 +188,20 @@ async function main() {
 		if (existsSync(sidecar)) unlinkSync(sidecar)
 	}
 
-	// Verify frozen artifact
-	const frozen = new DatabaseSync(outputPath, { readOnly: true })
-	const frozenMode = frozen.prepare("PRAGMA journal_mode").get() as { journal_mode: string }
-	frozen.close()
-
-	if (frozenMode.journal_mode !== "delete") {
-		throw new Error(`frozen DB journal_mode=${frozenMode.journal_mode}`)
-	}
-
 	// The sealed-artifact invariant: a built DB is a read-only asset from the moment it exists.
-	sealDatabase(outputPath)
+	sealDatabase(args.outputPath)
 	console.error("  sealed 0444")
 
-	const finalSize = (statSync(outputPath).size / 1024 / 1024).toFixed(1)
+	const finalSize = (statSync(args.outputPath).size / 1024 / 1024).toFixed(1)
 	const totalElapsed = ((performance.now() - t0) / 1000).toFixed(1)
 	console.error(`\nDone in ${totalElapsed}s:`)
-	console.error(`  Places:  ${processed.toLocaleString()}`)
+	console.error(`  Places:  ${ingest.placesIngested.toLocaleString()}`)
 
 	if (overtureIngested > 0) console.error(`  Overture: ${overtureIngested.toLocaleString()}`)
-	console.error(`  Skipped: ${skipped.toLocaleString()}`)
-	console.error(`  Output:  ${outputPath} (${finalSize} MB)`)
+	console.error(`  Skipped: ${ingest.skipped.toLocaleString()}`)
+	console.error(`  Output:  ${args.outputPath} (${finalSize} MB)`)
 }
 
-// Only run the full build when executed directly — importing this module (e.g. to reuse
-// `ingestOvertureDivisions` for an incremental single-country augment) must not trigger a build.
 if (import.meta.main) {
 	main().catch((err) => {
 		console.error(err)
